Add files via upload
Prediction of Mechanism of Action of a drug.

Abstract. The term "mechanism of action" is a pharmacological term commonly used when discussing medications or drugs. It refers to how the drug works on a molecular level in the body. The mechanism of action is an important aspect of drug development. A medicine's molecular mechanism of action links basic molecular interactions to reaction. The drug target, as well as the kinetic and conformational aspects that contribute to the communication of the desired response, are all important aspects of drug action for communicating a healthy, therapeutically useful response. This paper provides machine learning model to predict mechanism of action of a drug. We have also integrated this machine learning model into a website interface using Flask framework. A user can upload a custom testing features dataset, which contains the gene expression and cell viability levels. The web interface will output the top classes of drugs, along with the scatter plots for each of the drug. 

A web application using Flask framework. It uses Jinga templating engine.
- The input is a CSV file, which contains gene expression and cell viability values of each drug. 
- The output is the top classes of drugs (with the highest probability). 
- The machine learning model is serialized and de-serialized using the built-in keras ```save``` and ```load_model``` 
  (You can refer: https://keras.io/getting_started/faq/#what-are-my-options-for-saving-models)
Our dataset is divided into features dataset and target dataset. The training dataset consists of 23814 training samples. The testing dataset consists of 3982 testing samples.

A web application using Flask framework. It uses Jinga templating engine.
- The input is a CSV file, which contains gene expression and cell viability values of each drug. 
- The output is the top classes of drugs (with the highest probability). 
- The machine learning model is serialized and de-serialized using the built-in keras ```save``` and ```load_model``` 
  (You can refer: https://keras.io/getting_started/faq/#what-are-my-options-for-saving-models).